Data Mart Analysis
Data Mart is the latest venture and to analyze the sales and performance of my venture. In June 2020 - large scale supply changes were made at Data Mart. All Data Mart products now use sustainable packaging methods in every single step from the farm all the way to the customer.
To quantify the impact of this change on the sales performance for Data Mart and its separate business areas.